% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{apply_marker_settings}
\alias{apply_marker_settings}
\title{Apply marker settings to a pedigree}
\usage{
apply_marker_settings(x, mst)
}
\arguments{
\item{x}{a \code{\link[pedtools]{ped}} object or a list of such}

\item{mst}{a \code{data.frame} with the structure of a marker settings table}
}
\value{
an updated version of \code{x} to reflect the applied settings
}
\description{
Apply marker settings to a pedigree
}
