% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{custom_ped_labels}
\alias{custom_ped_labels}
\title{Get pedigree labels}
\usage{
custom_ped_labels(x)
}
\arguments{
\item{x}{a \code{\link[pedtools]{ped}} object or a list of such}
}
\value{
a character vector containing the ID labels of all pedigree members
}
\description{
Same as \code{\link[pedtools]{relabel}} but works on pedigree lists
}
\seealso{
\code{\link[pedtools]{relabel}}
}
