% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{custom_ped_plot}
\alias{custom_ped_plot}
\title{Plot a pedigree (list) for the forrel GUI}
\usage{
custom_ped_plot(x, available = c())
}
\arguments{
\item{x}{a \code{\link[pedtools]{ped}} object or a list of such}

\item{available}{a list of individuals available for genotyping (will be red)}
}
\description{
Plot a pedigree (list) for the forrel GUI
}
\seealso{
\code{\link[pedtools]{plotPedList}}
}
