% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{custom_ped_set_markers}
\alias{custom_ped_set_markers}
\title{Attach markers metadata to pedigree}
\usage{
custom_ped_set_markers(x, freqt)
}
\arguments{
\item{x}{a \code{\link[pedtools]{ped}} object or a list of such}

\item{freqt}{a \code{\link[fafreqs]{freqt}} object}
}
\value{
the pedigree (or pedList) with the specified markers attached
}
\description{
Attach markers metadata to pedigree
}
