% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{custom_read_ped}
\alias{custom_read_ped}
\title{Read a pedigree from a file for the forrel GUI}
\usage{
custom_read_ped(filepath)
}
\arguments{
\item{filepath}{the path to the pedigree file}
}
\value{
a \code{\link[pedtools]{ped}} object or a list of such
}
\description{
Wraps a \code{\link[pedtools]{readPed}} call so that invalid pedigree files
don't make shiny crash.
}
\seealso{
\code{\link[pedtools]{readPed}}
}
