% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{get_genotyped_ids}
\alias{get_genotyped_ids}
\title{Get genotyped individuals in a pedigree}
\usage{
get_genotyped_ids(x)
}
\arguments{
\item{x}{a \code{\link[pedtools]{ped}} object or a list of such}
}
\value{
a character vector containing the ID labels of all genotyped pedigree
  members
}
\description{
Get genotyped individuals in a pedigree
}
