% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{get_marker_names}
\alias{get_marker_names}
\title{Get a list of marker names from a pedigree}
\usage{
get_marker_names(x)
}
\arguments{
\item{x}{a \code{\link[pedtools]{ped}} object or a list of such}
}
\value{
a character vector containing the names of the markers attached to
  \code{x}. If \code{x} is a list of pedigrees, only the markers attached to
  the first component are used.
}
\description{
Get a list of marker names from a pedigree
}
