% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{read_familias_case_data}
\alias{read_familias_case_data}
\title{Import Familias case data}
\usage{
read_familias_case_data(x, file)
}
\arguments{
\item{x}{a \code{\link[pedtools]{ped}} object or a list of such}

\item{file}{a path to a Familias case data file}
}
\value{
an updated version of \code{x} with provided genotypes attached
}
\description{
Import Familias case data
}
