% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{read_generic_case_data}
\alias{read_generic_case_data}
\title{Read case data (in the generic format)}
\usage{
read_generic_case_data(x, file, ...)
}
\arguments{
\item{x}{a \code{\link[pedtools]{ped}} object or a list of such}

\item{file}{filename of an appropriate CSV file}

\item{...}{further parameters passed to \code{read.table}}
}
\value{
an updated version of \code{x} with the provided genotypes attached
}
\description{
Read case data (in the generic format)
}
