% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{rtrim}
\alias{rtrim}
\title{Trim a string from the right}
\usage{
rtrim(xs, nchars)
}
\arguments{
\item{xs}{a string or character vector}

\item{nchars}{the number of characters to remove (must be non-negative)}
}
\description{
Trim a string from the right
}
\examples{
rtrim("hello", 2) == "hel"

rtrim("hello", 10) == ""

rtrim("hello", 0) == "hello"

rtrim(c("hello", "bye"), 1) == c("hell", "by")
}
