% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{get_marker_settings_table}
\alias{get_marker_settings_table}
\title{Generate marker settings table from a pedigree}
\usage{
get_marker_settings_table(x, included_markers, mutation_settings)
}
\arguments{
\item{x}{a \code{\link[pedtools]{ped}} object or a list of such. If given a
list it only considers marker data in the first component.}

\item{included_markers}{a list of markers which will be included in the calculation}

\item{mutation_settings}{a list of mutation models corresponding to each marker}
}
\value{
a \code{data.frame} with the columns Marker, Mutations, Sex-linked?
  and Include in calculation? derived from the marker metadata in the
  pedigree
}
\description{
Generate marker settings table from a pedigree
}
